mapping ric steps
