+++ config/splat.pspeu.bobo6.yaml
@@ -1,0 +1,69 @@
+options:
+  platform: psp
+  basename: bobo6
+  base_path: ..
+  build_path: build/pspeu
+  target_path: disks/pspeu/PSP_GAME/USRDIR/res/ps/PSPBIN/bo6.bin
+  asm_path: asm/pspeu/boss/bo6_psp
+  asset_path: assets/boss/bo6
+  src_path: src/boss/bo6
+  ld_script_path: build/pspeu/bobo6.ld
+  compiler: GCC
+  symbol_addrs_path:
+    - config/symbols.pspeu.txt
+    - config/symbols.pspeu.bobo6.txt
+  undefined_funcs_auto_path: config/undefined_funcs_auto.pspeu.bobo6.txt
+  undefined_syms_auto_path: config/undefined_syms_auto.pspeu.bobo6.txt
+  find_file_boundaries: true
+  use_legacy_include_asm: false
+  migrate_rodata_to_functions: true
+  asm_jtbl_label_macro: jlabel
+  symbol_name_format: pspeu_$VRAM
+  section_order:
+    - .text
+    - .data
+    - .rodata
+    - .bss
+  ld_bss_is_noload: true
+  disasm_unknown: true
+  disassemble_all: true
+  global_vram_start: 0x8000000
+  asm_inc_header: |
+    .set noat      /* allow manual use of $at */
+    .set noreorder /* don't insert nops after branches */
+sha1: 544630b046da708572014835760914c387cbc5dd
+segments:
+  - [0x0, bin, mwo_header]
+  - name: bobo6
+    type: code
+    start: 0x80
+    vram:  0x09237700
+    bss_size: 0x2A80
+    align: 128
+    subalign: 8
+    subsegments:
+      - [0x80, c, bo6_psp/80] # create_entity
+      # - [0x1028, c, bo6_psp/80] # e_stage_name
+      # - [0x1EA8, c, bo6_psp/80] # e_red_door
+      # - [0x31B0, c, bo6_psp/80] # st_update
+      # - [0x36E8, c, bo6_psp/80] # e_particles
+      # - [0x4C78, c, bo6_psp/80]
+      # - [0x8EC0, c, bo6_psp/80] # e_misc
+      # - [0x9470, c, bo6_psp/80] # collision
+      # - [0xB648, c, bo6_psp/80] # popup
+      # - [0xBC40, c, bo6_psp/80] # e_room_fg
+      # - [0xBD58, c, bo6_psp/80] # st_common
+      # - [0xC840, c, bo6_psp/80] # prim_helpers
+      # - [0xCF80, c, bo6_psp/80]
+      # - [0xD380, c, bo6_psp/80] # e_breakable
+      # - [0xD500, c, bo6_psp/80]
+      # - [0x10340, c, bo6_psp/80] # e_room_bg
+      # - [0x10420, c, bo6_psp/80]
+      # - [0x11248, c, bo6_psp/80] # e_lock_camera
+      # - [0x114A8, c, bo6_psp/80]
+      # - [0x23538, c, bo6_psp/80] # pl_utils
+      # - [0x24000, c, bo6_psp/80]
+      - [0x33B80, data]
+      - [0x43F80, .rodata, 80]
+      - {type: bss, vram: 0x927B980}
+  - [0x44380]
