+++ config/symbols.pspeu.bobo6.txt
@@ -1,0 +1,66 @@
+CreateEntityFromLayout = 0x09237700; // source=match
+CreateEntityWhenInVerticalRange = 0x09237820; // source=match
+CreateEntityWhenInHorizontalRange = 0x092379B8; // source=match
+FindFirstEntityToTheRight = 0x09237B50; // source=match
+FindFirstEntityToTheLeft = 0x09237BC0; // source=match
+CreateEntitiesToTheRight = 0x09237C48; // source=match
+CreateEntitiesToTheLeft = 0x09237D88; // source=match
+FindFirstEntityAbove = 0x09237EE0; // source=match
+FindFirstEntityBelow = 0x09237F58; // source=match
+CreateEntitiesAbove = 0x09237FE0; // source=match
+CreateEntitiesBelow = 0x09238128; // source=match
+InitRoomEntities = 0x09238288; // source=match
+UpdateRoomPosition = 0x092384B8; // source=match
+CreateEntityFromCurrentEntity = 0x09238598; // source=match
+CreateEntityFromEntity = 0x09238620; // source=match
+EntityStageNamePopup = 0x092386A8; // source=match
+EntityIsNearPlayer = 0x09239528; // source=match
+BO6_EntityRedDoor = 0x09239610; // source=match
+Random = 0x0923A830; // source=match
+Update = 0x0923A870; // source=match
+UpdateStageEntities = 0x0923AC50; // source=match
+func_psp_0923AD68 = 0x0923AD68; // source=match
+func_psp_0923B2F0 = 0x0923B2F0; // source=match
+EntitySoulStealOrb = 0x0923B780; // source=match
+EntityEnemyBlood = 0x0923BCE0; // source=match
+EntityUnkId13 = 0x09240540; // source=match
+EntityExplosionVariants = 0x09240688; // source=match
+EntityGreyPuff = 0x09240800; // source=match
+EntityIntenseExplosion = 0x09240980; // source=match
+HitDetection = 0x09240AF0; // source=match
+EntityDamageDisplay = 0x092422C0; // source=match
+BottomCornerText = 0x09242CC8; // source=match
+EntityRoomForeground = 0x092432C0; // source=match
+DestroyEntity = 0x092433D8; // source=match
+PreventEntityFromRespawning = 0x09243470; // source=match
+AnimateEntity = 0x092434F8; // source=match
+GetSideToPlayer = 0x09243668; // source=match
+MoveEntity = 0x092436F0; // source=match
+FallEntity = 0x09243740; // source=match
+AllocEntity = 0x09243778; // source=match
+UnkEntityFunc0 = 0x092437E8; // source=match
+GetAngleBetweenEntities = 0x09243890; // source=match
+GetNormalizedAngle = 0x09243910; // source=match
+SetStep = 0x09243A08; // source=match
+InitializeEntity = 0x09243A58; // source=match
+EntityDummy = 0x09243C00; // source=match
+CheckFieldCollision = 0x09243C38; // source=match
+ReplaceBreakableWithItemDrop = 0x09243DC8; // source=match
+UnkPrimHelper = 0x09243EC0; // source=match
+PrimResetNext = 0x092443C0; // source=match
+UnkPolyFunc2 = 0x092444C8; // source=match
+PrimDecreaseBrightness = 0x09244530; // source=match
+EntityBreakable = 0x09244A00; // source=match
+BO6_EntityBackgroundBlock = 0x092479C0; // source=match
+BO6_EntityLockCamera = 0x092488C8; // source=match
+func_8015C4AC = 0x0925ABB8; // source=match
+func_8015C6D4 = 0x0925AE20; // source=match
+RicSetStep = 0x0925B248; // source=match
+RicSetAnimation = 0x0925B270; // source=match
+RicDecelerateX = 0x0925B2B0; // source=match
+RicCheckFacing = 0x0925B350; // source=match
+RicSetSpeedX = 0x0925B448; // source=match
+func_8015CAAC = 0x0925B498; // source=match
+RicSetInvincibilityFrames = 0x0925B4E0; // source=match
+DisableAfterImage = 0x0925B588; // source=match
+func_8015CC28 = 0x0925B648; // source=match
