+++ config/symbols.us.bobo6.txt
@@ -339,5 +339,26 @@ g_Ric_prev_step_s = 0x801D166A;
 g_Ric_unk5C = 0x801D1674;
 g_Ric_unk6A = 0x801D1682;
 g_Ric_unk6C = 0x801D1684;
-g_Ric_unk70 = 0x801D1688;
+e_Ric_unk70 = 0x801D1688;
 g_ItemIconSlots = 0x801D16A8;
+BO6_DisableAfterImage = 0x801B9B78;
+BO6_RicSetInvincibilityFrames = 0x801B9AF4;
+BO6_RicSetStep = 0x801B9928;
+BO6_RicStepStand = 0x801B6BD8;
+BO6_RicStepWalk = 0x801B6E3C;
+BO6_RicStepCrouch = 0x801B74DC;
+BO6_RicStepFall = 0x801B7444;
+BO6_RicStepJump = 0x801B6F88;
+BO6_RicStepHighJump = 0x801B8F78;
+BO6_RicStepHit = 0x801B7818;
+BO6_RicStepDead = 0x801B7E80;
+BO6_RicStepStandInAir = 0x801B8540;
+BO6_RicStepEnableFlameWhip = 0x801B85DC;
+BO6_RicStepHydrostorm = 0x801B8684;
+BO6_RicStepThrowDaggers = 0x801B86EC;
+BO6_RicStepGenericSubwpnCrash = 0x801B86B8;
+BO6_RicStepSlide = 0x801B87B4;
+BO6_RicStepRun = 0x801B6EA8;
+BO6_RicStepSlideKick = 0x801B8A88;
+BO6_RicStepBladeDash = 0x801B8D50;
+BO6_RicMain = 0x801B50BC;
